# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NIDAN.ai API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface CreateGeminiConversationBody {
  title: string;
}

export interface SendGeminiMessageBody {
  content: string;
}

export type ReportType = (typeof ReportType)[keyof typeof ReportType];

export const ReportType = {
  image: "image",
  ocr: "ocr",
} as const;

export interface Report {
  id: number;
  userId: string;
  type: ReportType;
  filename: string;
  imageBase64?: string;
  analysisResult: string;
  findings?: string;
  recommendations?: string;
  exportCount: number;
  createdAt: string;
}

export type AnalyzeReportBodyType =
  (typeof AnalyzeReportBodyType)[keyof typeof AnalyzeReportBodyType];

export const AnalyzeReportBodyType = {
  image: "image",
  ocr: "ocr",
} as const;

export interface AnalyzeReportBody {
  imageBase64: string;
  filename: string;
  type: AnalyzeReportBodyType;
  userId: string;
}

export interface Doctor {
  id: number;
  name: string;
  specialty: string;
  hospital: string;
  rating: number;
  experience: number;
  location: string;
  phone?: string;
  available: boolean;
}

export interface DashboardStats {
  totalConsultations: number;
  imagesAnalyzed: number;
  reportsExported: number;
  activeConversations: number;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  chat: "chat",
  image_analysis: "image_analysis",
  export: "export",
  ocr: "ocr",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  timestamp: string;
}

export interface ApiError {
  error: string;
}

export type ListReportsParams = {
  type?: ListReportsType;
};

export type ListReportsType =
  (typeof ListReportsType)[keyof typeof ListReportsType];

export const ListReportsType = {
  image: "image",
  ocr: "ocr",
} as const;

export type ListDoctorsParams = {
  specialty?: string;
  query?: string;
};