# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NIDAN.ai API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: gemini
    description: Gemini AI chat and image operations
  - name: reports
    description: Medical image analysis and reports
  - name: doctors
    description: Doctor recommendations
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeminiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"

  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive an AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendGeminiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}

  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all medical image analysis reports
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [image, ocr]
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"

  /reports/analyze:
    post:
      operationId: analyzeReport
      tags: [reports]
      summary: Upload and analyze a medical image or document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeReportBody"
      responses:
        "201":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"

  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a specific report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteReport
      tags: [reports]
      summary: Delete a report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      summary: List recommended doctors based on specialty
      parameters:
        - name: specialty
          in: query
          required: false
          schema:
            type: string
        - name: query
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of doctors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Doctor"

  /stats:
    get:
      operationId: getDashboardStats
      tags: [stats]
      summary: Get dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /stats/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [stats]
      summary: Get recent user activity feed
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    GeminiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt

    GeminiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    GeminiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"
      required:
        - id
        - title
        - createdAt
        - messages

    CreateGeminiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    SendGeminiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    Report:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
          enum: [image, ocr]
        filename:
          type: string
        imageBase64:
          type: string
        analysisResult:
          type: string
        findings:
          type: string
        recommendations:
          type: string
        exportCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - filename
        - analysisResult
        - exportCount
        - createdAt

    AnalyzeReportBody:
      type: object
      properties:
        imageBase64:
          type: string
        filename:
          type: string
        type:
          type: string
          enum: [image, ocr]
        userId:
          type: string
      required:
        - imageBase64
        - filename
        - type
        - userId

    Doctor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        specialty:
          type: string
        hospital:
          type: string
        rating:
          type: number
        experience:
          type: integer
        location:
          type: string
        phone:
          type: string
        available:
          type: boolean
      required:
        - id
        - name
        - specialty
        - hospital
        - rating
        - experience
        - location
        - available

    DashboardStats:
      type: object
      properties:
        totalConsultations:
          type: integer
        imagesAnalyzed:
          type: integer
        reportsExported:
          type: integer
        activeConversations:
          type: integer
      required:
        - totalConsultations
        - imagesAnalyzed
        - reportsExported
        - activeConversations

    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [chat, image_analysis, export, ocr]
        description:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - type
        - description
        - timestamp

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error