# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NIDAN.ai API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all conversations
 */
export const ListGeminiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGeminiConversationsResponse = zod.array(
  ListGeminiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateGeminiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetGeminiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.coerce.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteGeminiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListGeminiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListGeminiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListGeminiMessagesResponse = zod.array(
  ListGeminiMessagesResponseItem,
);

/**
 * @summary Send a message and receive an AI response (SSE stream)
 */
export const SendGeminiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendGeminiMessageBody = zod.object({
  content: zod.string(),
});

/**
 * @summary List all medical image analysis reports
 */
export const ListReportsQueryParams = zod.object({
  type: zod.enum(["image", "ocr"]).optional(),
});

export const ListReportsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  type: zod.enum(["image", "ocr"]),
  filename: zod.string(),
  imageBase64: zod.string().optional(),
  analysisResult: zod.string(),
  findings: zod.string().optional(),
  recommendations: zod.string().optional(),
  exportCount: zod.number(),
  createdAt: zod.coerce.date(),
});
export const ListReportsResponse = zod.array(ListReportsResponseItem);

/**
 * @summary Upload and analyze a medical image or document
 */
export const AnalyzeReportBody = zod.object({
  imageBase64: zod.string(),
  filename: zod.string(),
  type: zod.enum(["image", "ocr"]),
  userId: zod.string(),
});

/**
 * @summary Get a specific report
 */
export const GetReportParams = zod.object({
  id: zod.coerce.number(),
});

export const GetReportResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  type: zod.enum(["image", "ocr"]),
  filename: zod.string(),
  imageBase64: zod.string().optional(),
  analysisResult: zod.string(),
  findings: zod.string().optional(),
  recommendations: zod.string().optional(),
  exportCount: zod.number(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a report
 */
export const DeleteReportParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List recommended doctors based on specialty
 */
export const ListDoctorsQueryParams = zod.object({
  specialty: zod.coerce.string().optional(),
  query: zod.coerce.string().optional(),
});

export const ListDoctorsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  specialty: zod.string(),
  hospital: zod.string(),
  rating: zod.number(),
  experience: zod.number(),
  location: zod.string(),
  phone: zod.string().optional(),
  available: zod.boolean(),
});
export const ListDoctorsResponse = zod.array(ListDoctorsResponseItem);

/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalConsultations: zod.number(),
  imagesAnalyzed: zod.number(),
  reportsExported: zod.number(),
  activeConversations: zod.number(),
});

/**
 * @summary Get recent user activity feed
 */
export const GetRecentActivityResponseItem = zod.object({
  id: zod.number(),
  type: zod.enum(["chat", "image_analysis", "export", "ocr"]),
  description: zod.string(),
  timestamp: zod.coerce.date(),
});
export const GetRecentActivityResponse = zod.array(
  GetRecentActivityResponseItem,
);