# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NIDAN.ai API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalConsultations: number;
  imagesAnalyzed: number;
  reportsExported: number;
  activeConversations: number;
}