# lib/api-zod/src/generated/types/geminiMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NIDAN.ai API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: Date;
}