# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NIDAN.ai API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReportType } from "./reportType";

export interface Report {
  id: number;
  userId: string;
  type: ReportType;
  filename: string;
  imageBase64?: string;
  analysisResult: string;
  findings?: string;
  recommendations?: string;
  exportCount: number;
  createdAt: Date;
}